+++ frontend/src/components/inventory-ui/AddAsset.tsx
@@ -59,7 +59,7 @@ const AddAsset = ({ defaultValues }: { defaultValues: Defaults }) => {
 
   const hardwareDefaultValues = {
     ...defaultAssetValues,
-    category: defaultValues.hardwareCategory,
+    category: defaultValues.hardwareCategory ?? '',
     processor: '',
     memory: '',
     storage: '',
@@ -75,12 +75,12 @@ const AddAsset = ({ defaultValues }: { defaultValues: Defaults }) => {
   
   const softwareDefaultValues = {
     ...defaultAssetValues,
-    category: defaultValues.softwareCategory,
+    category: defaultValues.softwareCategory ?? '',
     vendor: '',
     licenseType: '',
     purchaseDate: undefined,
     licenseExpirationDate: undefined,
-    type: undefined,
+    type: 'Software' as AssetType,
     license: '',
     version: '',
     noOfLicense: '1',
@@ -177,8 +177,8 @@ const AddAsset = ({ defaultValues }: { defaultValues: Defaults }) => {
               <form className="flex flex-col w-full gap-4" onSubmit={form.handleSubmit(onSubmit)}>
                 <Tabs defaultValue="Hardware" className="w-full">
                   <TabsList className="grid w-full grid-cols-2">
-                    <TabsTrigger value="Hardware" onClick={() => handleTabChange("Hardware")}>Hardware</TabsTrigger>
-                    <TabsTrigger value="Software" onClick={() => handleTabChange("Software")}>Software</TabsTrigger>
+                    <TabsTrigger className='data-[state=active]:bg-tab' value="Hardware" onClick={() => handleTabChange("Hardware")}>Hardware</TabsTrigger>
+                    <TabsTrigger className='data-[state=active]:bg-tab' value="Software" onClick={() => handleTabChange("Software")}>Software</TabsTrigger>
                   </TabsList>
                   <div className="">
                     <TabsContent tabIndex={-1} value="Hardware" className="pb-4 px-3">